redo syntax such that erlang types are the only types that exist
It used to be that there where several structures for different types
in Joxa. These structures where annotated with thier types an the
jxa_annot object. However, those annotations would not be available to
macros and therefor macros would not be usable really at all. So te
simplify the ast to macro conversions for the macro writer the Joxa
typse where simplified to just be Erlang types without additional
annotations. In addition certain structures became both slightly more
complex and slightly less usable. for example, Literal lists are
created with the [] syntax. That works very well but that syntax is
now just a reader macro for (list <args>). That is [1 2 3 4] simply
gets translated to the AST (list 1 2 3 4).