support automatic inlining in joxa
This supports automatic inlining in joxa via the very nice core erlang
inline optimizer. In talks with Richard Carlsson I found out that this
is much much more then just an inliner. It is, in fact a very
aggressive optimizer in its own right. This patch enables it as part
of the final compilation on builds.