Fix issue with `:joxify` element in require
This makes sure that the return value of 'joxa-compiler':info/1,2
always contains a list of *all* module dependencies.

Prior to that, there was an issue with a module's name being missing
from the list of module dependencies if that particular module was
required with the `:joxify` element.

This adds not only a test case for require clauses with the `:joxify`
element, but also one for those with an `:as` element.