Fix issue with `:as` and `:joxify` in one require
In the previous commit 224c773c56, we fixed an issue with missing
module dependency information when a require clause contained a
`:joxify` element. In fact, the bugfix was not complete, as it didn't
cover the case with both an `:as` and a `:joxify` element in a require
clause.

Retrieving namespace information for require clauses is now basically
done the same way like for use clauses, i.e., we don't check for the
number, order, or presence of particular elements.