Fix issue with `let/3` macro in records
This fixes an issue with unwanted variable capture in the `let/3` macro
that's generated by the `defrecord/2` macro. The issue occurred when
the name of the first variable passed to `let/3` (i.e., the record) and
the name of that very record would be the same.

In the corresponding test suite, I've renamed the record that's used
for testing from `my-records` to `test-record`, and renamed any of its
instances being used from `one` to `test-record` as well. Firstly, I
did that in order to have a failing test case for this fix here, but
also in order to cover similar potential issues with other macros that
are being tested there.